+++ src/features/notifications/api/index.ts
@@ -1,3 +1,2 @@
 export * from './notifications.api'
-export * from './notifications.socket'
 export * from './notifications.type'
